refactor: share request classification
