+++ crates/mesh-llm-host-runtime/src/network/router.rs
@@ -1,408 +1,8 @@
 /// Smart model router — classifies requests and picks the best model.
-use serde_json::Value;
-
-// ── Request categories ──────────────────────────────────────────────
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Category {
-    Code,
-    Reasoning,
-    Chat,
-    ToolCall,
-    Creative,
-    /// Factual lookup, summarization, knowledge retrieval
-    Info,
-    /// Image generation or analysis (future: multimodal models)
-    Image,
-}
-
-/// How complex/heavy the request appears to be.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Complexity {
-    Quick,    // simple fact, short answer, casual
-    Moderate, // normal conversation, standard code
-    Deep,     // long reasoning, complex analysis, architecture
-}
-
-/// Full classification result.
-#[derive(Debug, Clone, PartialEq)]
-pub struct Classification {
-    pub category: Category,
-    pub complexity: Complexity,
-    pub needs_tools: bool,
-    pub has_media_inputs: bool,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct MediaRequirements {
-    pub has_media: bool,
-    pub needs_vision: bool,
-    pub needs_audio: bool,
-}
-
-impl MediaRequirements {
-    pub fn requires_runtime_modality(self) -> bool {
-        self.needs_vision || self.needs_audio
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-struct RouterSignalScores {
-    code: usize,
-    reasoning: usize,
-    creative: usize,
-    info: usize,
-    image: usize,
-    deep: usize,
-}
-
-// ── Model capabilities for routing ──────────────────────────────────
-
-/// Strip split GGUF suffix like "-00001-of-00004" from a model name.
-pub fn strip_split_suffix(name: &str) -> &str {
-    // Pattern: -NNNNN-of-NNNNN at the end
-    if let Some(idx) = name.rfind("-of-") {
-        // Check that what follows is digits and what precedes is -digits
-        let after = &name[idx + 4..];
-        if after.chars().all(|c| c.is_ascii_digit()) && !after.is_empty() {
-            // Find the preceding -NNNNN
-            if let Some(dash) = name[..idx].rfind('-') {
-                let between = &name[dash + 1..idx];
-                if between.chars().all(|c| c.is_ascii_digit()) && !between.is_empty() {
-                    return &name[..dash];
-                }
-            }
-        }
-    }
-    name
-}
-
-/// Owned version of strip_split_suffix for contexts that need a String.
-pub fn strip_split_suffix_owned(name: &str) -> String {
-    strip_split_suffix(name).to_string()
-}
-
-// ── Request classification ──────────────────────────────────────────
-
-/// Classify a chat completion request body using heuristics.
-/// No LLM call, just pattern matching on the request structure.
-/// Classify a request body into category + complexity + needs_tools.
-/// Tools presence is an attribute, not a category override — a code request
-/// with tools is still Code (with needs_tools=true), not ToolCall.
-pub fn classify(body: &Value) -> Classification {
-    let text = collect_message_text(body);
-    let lower = text.to_lowercase();
-    let media = media_requirements(body);
-    let needs_tools = detect_tool_requirement(body);
-    let last_user_len = last_user_message_len(body);
-    let scores = router_signal_scores(&lower);
-    let category = classify_category(scores, detect_system_code_hint(body), media, needs_tools);
-    let complexity = classify_complexity(scores, last_user_len, message_count(body));
-
-    Classification {
-        category,
-        complexity,
-        needs_tools,
-        has_media_inputs: media.has_media,
-    }
-}
-
-fn detect_tool_requirement(body: &Value) -> bool {
-    has_tools_schema(body) || has_tool_blocks(body)
-}
-
-fn has_tools_schema(body: &Value) -> bool {
-    body.get("tools")
-        .and_then(|t| t.as_array())
-        .map(|a| !a.is_empty())
-        .unwrap_or(false)
-}
-
-fn has_tool_blocks(body: &Value) -> bool {
-    body.get("messages")
-        .and_then(|m| m.as_array())
-        .map(|msgs| {
-            msgs.iter().any(|msg| {
-                msg.get("content")
-                    .and_then(|c| c.as_array())
-                    .map(|blocks| {
-                        blocks.iter().any(|b| {
-                            matches!(
-                                b.get("type").and_then(|t| t.as_str()),
-                                Some("tool_use") | Some("tool_result")
-                            )
-                        })
-                    })
-                    .unwrap_or(false)
-            })
-        })
-        .unwrap_or(false)
-}
-
-fn count_signals(lower: &str, signals: &[&str]) -> usize {
-    signals
-        .iter()
-        .filter(|signal| lower.contains(*signal))
-        .count()
-}
-
-fn router_signal_scores(lower: &str) -> RouterSignalScores {
-    RouterSignalScores {
-        code: count_signals(
-            lower,
-            &[
-                "```",
-                "def ",
-                "fn ",
-                "func ",
-                "class ",
-                "import ",
-                "function",
-                "const ",
-                "let ",
-                "var ",
-                "return ",
-                "write a program",
-                "write code",
-                "implement",
-                "refactor",
-                "debug",
-                "fix the bug",
-                "write a script",
-                "code review",
-                "pull request",
-                "git ",
-                "compile",
-                "syntax",
-                "python",
-                "javascript",
-                "typescript",
-                " rust ",
-                "golang",
-                "java ",
-                "c++",
-                " ruby ",
-                " swift ",
-                "kotlin",
-                "algorithm",
-                "binary search",
-                " sort ",
-                "regex",
-                " api ",
-                " http ",
-                " sql ",
-                "database",
-                " query ",
-            ],
-        ),
-        reasoning: count_signals(
-            lower,
-            &[
-                "prove",
-                "explain why",
-                "step by step",
-                "calculate",
-                "solve",
-                "derive",
-                "what is the probability",
-                "how many",
-                "analyze",
-                "compare and contrast",
-                "evaluate",
-                "mathematical",
-                "theorem",
-                "equation",
-                "logic",
-                "think carefully",
-                "reason about",
-            ],
-        ),
-        creative: count_signals(
-            lower,
-            &[
-                "write a story",
-                "write a poem",
-                "creative",
-                "imagine",
-                "fiction",
-                "narrative",
-                "compose",
-                "brainstorm",
-                "write a song",
-                "screenplay",
-                "dialogue",
-            ],
-        ),
-        info: count_signals(
-            lower,
-            &[
-                "what is",
-                "who is",
-                "when did",
-                "where is",
-                "how does",
-                "define ",
-                "explain ",
-                "summarize",
-                "summary",
-                "overview",
-                "tell me about",
-                "describe ",
-                "what are the",
-                "list the",
-                "difference between",
-                "compare ",
-                "history of",
-            ],
-        ),
-        image: count_signals(
-            lower,
-            &[
-                "image",
-                "picture",
-                "photo",
-                "draw",
-                "generate an image",
-                "visualize",
-                "diagram",
-                "screenshot",
-                "describe this image",
-            ],
-        ),
-        deep: count_signals(
-            lower,
-            &[
-                "architect",
-                "design a system",
-                "trade-off",
-                "tradeoff",
-                "in depth",
-                "comprehensive",
-                "thorough",
-                "detailed analysis",
-                "long-term",
-                "strategy",
-                "plan for",
-                "review this codebase",
-                "rewrite",
-                "from scratch",
-            ],
-        ),
-    }
-}
-
-fn detect_system_code_hint(body: &Value) -> bool {
-    body.get("messages")
-        .and_then(|m| m.as_array())
-        .map(|messages| {
-            messages.iter().any(|msg| {
-                msg.get("role").and_then(|r| r.as_str()) == Some("system")
-                    && msg
-                        .get("content")
-                        .and_then(|c| c.as_str())
-                        .map(|content| {
-                            let sys = content.to_lowercase();
-                            sys.contains("developer")
-                                || sys.contains("coding")
-                                || sys.contains("programmer")
-                        })
-                        .unwrap_or(false)
-            })
-        })
-        .unwrap_or(false)
-}
-
-fn classify_category(
-    scores: RouterSignalScores,
-    system_code: bool,
-    media: MediaRequirements,
-    needs_tools: bool,
-) -> Category {
-    if system_code
-        || scores.code >= 2
-        || (scores.code >= 1 && scores.reasoning == 0 && scores.creative == 0)
-    {
-        Category::Code
-    } else if scores.reasoning >= 2 {
-        Category::Reasoning
-    } else if scores.creative >= 1 {
-        Category::Creative
-    } else if media.needs_vision || scores.image >= 1 {
-        Category::Image
-    } else if needs_tools && scores.code == 0 && scores.reasoning == 0 && scores.creative == 0 {
-        Category::ToolCall
-    } else if scores.info >= 2 && scores.code == 0 {
-        Category::Info
-    } else {
-        Category::Chat
-    }
-}
-
-fn message_count(body: &Value) -> usize {
-    body.get("messages")
-        .and_then(|m| m.as_array())
-        .map(|a| a.len())
-        .unwrap_or(0)
-}
-
-fn classify_complexity(
-    scores: RouterSignalScores,
-    last_user_len: usize,
-    total_messages: usize,
-) -> Complexity {
-    if scores.deep >= 1 || last_user_len > 500 || total_messages > 10 {
-        Complexity::Deep
-    } else if last_user_len < 60 && total_messages <= 2 && scores.reasoning == 0 && scores.deep == 0
-    {
-        Complexity::Quick
-    } else {
-        Complexity::Moderate
-    }
-}
-
-pub fn media_requirements(body: &Value) -> MediaRequirements {
-    let mut requirements = MediaRequirements::default();
-    let Some(messages) = body.get("messages").and_then(|m| m.as_array()) else {
-        return requirements;
-    };
-
-    for msg in messages {
-        let Some(blocks) = msg.get("content").and_then(|c| c.as_array()) else {
-            continue;
-        };
-        for block in blocks {
-            let block_type = block
-                .get("type")
-                .and_then(|t| t.as_str())
-                .unwrap_or_default();
-            match block_type {
-                "image_url" | "input_image" | "image" => {
-                    requirements.has_media = true;
-                    requirements.needs_vision = true;
-                }
-                "audio_url" | "input_audio" | "audio" => {
-                    requirements.has_media = true;
-                    requirements.needs_audio = true;
-                }
-                "file" | "input_file" => {
-                    requirements.has_media = true;
-                }
-                _ => {
-                    if block.get("image_url").is_some() || block.get("image").is_some() {
-                        requirements.has_media = true;
-                        requirements.needs_vision = true;
-                    }
-                    if block.get("audio_url").is_some() || block.get("audio").is_some() {
-                        requirements.has_media = true;
-                        requirements.needs_audio = true;
-                    }
-                }
-            }
-        }
-    }
-
-    requirements
-}
+pub use mesh_llm_types::router::{
+    Category, Classification, Complexity, MediaRequirements, classify, media_requirements,
+    strip_split_suffix, strip_split_suffix_owned,
+};
 
 pub(crate) fn model_satisfies_media_requirements(
     caps: &crate::models::ModelCapabilities,
@@ -428,55 +28,6 @@ pub(crate) fn filter_media_compatible_candidates<'a>(
     } else {
         Some(media_available)
     }
-}
-
-/// Length of last user message in characters (rough complexity proxy).
-fn last_user_message_len(body: &Value) -> usize {
-    body.get("messages")
-        .and_then(|m| m.as_array())
-        .and_then(|msgs| {
-            msgs.iter()
-                .rev()
-                .find(|m| m.get("role").and_then(|r| r.as_str()) == Some("user"))
-        })
-        .map(message_text)
-        .map(|s| s.len())
-        .unwrap_or(0)
-}
-
-fn collect_message_text(body: &Value) -> String {
-    let mut text = String::new();
-    if let Some(messages) = body.get("messages").and_then(|m| m.as_array()) {
-        for msg in messages {
-            let content = message_text(msg);
-            if !content.is_empty() {
-                text.push_str(&content);
-                text.push('\n');
-            }
-        }
-    }
-    text
-}
-
-/// Extract message text for both OpenAI-style and Anthropic-style payloads.
-fn message_text(msg: &Value) -> String {
-    if let Some(s) = msg.get("content").and_then(|c| c.as_str()) {
-        return s.to_string();
-    }
-
-    // Anthropic content blocks: [{"type":"text","text":"..."}, ...]
-    if let Some(blocks) = msg.get("content").and_then(|c| c.as_array()) {
-        let mut out = String::new();
-        for b in blocks {
-            if let Some(t) = b.get("text").and_then(|t| t.as_str()) {
-                out.push_str(t);
-                out.push('\n');
-            }
-        }
-        return out;
-    }
-
-    String::new()
 }
 
 // ── Model selection ─────────────────────────────────────────────────
